Improve AltGr modifier triggering detection
A press/release of AltGr is translated by the system into the press/release of two keys - LCtrl and RAlt. In Far, for modifiers, activation is registered upon release, and the problem with AltGr is that, being a combined event, it triggered upon the release of the very first key in the combination - LCtrl, thereby producing "CtrlRAlt" instead of the expected "RAlt" (similarly for compound combinations).

This patch prevents premature reaction to the LCtrl release event.